Create poker-game.py
This program simulates a game of poker. I have added a provision for resolving ties, and a bunch of test cases towards the end for debugging purposes.
The function 'hand_percentages' is purely utilitarian. It is used to basically simulate a bunch of "winning" poker hands, and compute their probability to a reasonable approximation.